+++ apps/example/app.config.js
@@ -9,7 +9,20 @@ module.exports = {
     "userInterfaceStyle": "automatic",
     "ios": {
       "supportsTablet": true,
-      "bundleIdentifier": "com.nitrostorage.example"
+      "bundleIdentifier": "com.nitrostorage.example",
+      "infoPlist": {
+        "UIApplicationSceneManifest": {
+          "UIApplicationSupportsMultipleScenes": false,
+          "UISceneConfigurations": {
+            "UIWindowSceneSessionRoleApplication": [
+              {
+                "UISceneConfigurationName": "Default Configuration",
+                "UISceneDelegateClassName": "$(PRODUCT_MODULE_NAME).SceneDelegate"
+              }
+            ]
+          }
+        }
+      }
     },
     "android": {
       "package": "com.nitrostorage.example",
@@ -20,6 +33,7 @@ module.exports = {
     },
     "plugins": [
       "expo-router",
+      "./plugins/with-ios-scene-lifecycle",
       [
         "expo-build-properties",
         {